feat: moved networkHandler to fb common utils
